backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
pydantic[email]
pytest
httpx
google-genai
python-dotenv
python-jose[cryptography]
bcrypt
